Function to create a crystallographic_ls class, not just an instance
The builder function crystallographic_ls is not enough for Olex2 because
one must be able to inherit from the smtbx class in Olex2, in order to
override key operations to send back results to Olex2.

The solution is therefore to create a class builder in addition to the
instance builder.